Possibilita criação de pedidos

+++ src/main/java/com/carolruo/projeto/domain/Product.java
@@ -15,7 +15,7 @@ public class Product implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
-    private BigDecimal price;
+    private Double price;
     //    @JsonIgnore
     @JsonIgnore
     @ManyToMany
@@ -31,7 +31,7 @@ public class Product implements Serializable {
     public Product() {
     }
 
-    public Product(Integer id, String name, BigDecimal price) {
+    public Product(Integer id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -62,11 +62,11 @@ public class Product implements Serializable {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
